moving files to contrib
