+++ src/lbm/contrib/data/aspect_ratios.py
