+++ src/lbm/data/mappers/__init__.py
@@ -1,12 +1,9 @@
 from .base import BaseMapper
-from .mappers import KeyRenameMapper, RescaleMapper, TorchvisionMapper, RandomPixelMasking, CustomResize, RandomMask
+from .mappers import KeyRenameMapper, RescaleMapper, TorchvisionMapper
 from .mappers_config import (
     KeyRenameMapperConfig,
     RescaleMapperConfig,
     TorchvisionMapperConfig,
-    RandomPixelMaskingConfig,
-    CustomResizeConfig,
-    RandomMaskConfig,
 )
 from .mappers_wrapper import MapperWrapper
 
@@ -18,11 +15,5 @@ __all__ = [
     "KeyRenameMapperConfig",
     "RescaleMapperConfig",
     "TorchvisionMapperConfig",
-    "RandomPixelMaskingConfig",
     "MapperWrapper",
-    "RandomPixelMasking",
-    "CustomResize",
-    "CustomResizeConfig",
-    "RandomMask",
-    "RandomMaskConfig"
 ]
